feat(octavia): added provioning and operating status form members
ref: MANAGER-12807
