+++ packages/manager/modules/octavia-load-balancer/src/load-balancers/load-balancer/listeners/component.js
@@ -1,10 +1,9 @@
-import controller from './controller';
 import template from './template.html';
 
 export default {
   bindings: {
     listeners: '<',
+    discoverLink: '<',
   },
-  controller,
   template,
 };
