--- packages/manager/modules/octavia-load-balancer/src/load-balancers/load-balancer/listeners/controller.js
@@ -1,7 +1,0 @@
-import { DISCOVER_LINK } from '../constants';
-
-export default class OctaviaLoadBalancerListenersCtrl {
-  constructor() {
-    this.link = DISCOVER_LINK;
-  }
-}
