+++ packages/manager/modules/octavia-load-balancer/src/load-balancers/load-balancer/listeners/list/component.js
@@ -10,6 +10,8 @@ export default {
     goToListenerDeletion: '<',
     getPoolDetailLink: '<',
     trackRoot: '<',
+    operatingStatusBadges: '<',
+    provisioningStatusBadges: '<',
   },
   controller,
   template,
