+++ packages/manager/modules/octavia-load-balancer/src/load-balancers/load-balancer/listeners/list/controller.js
@@ -2,16 +2,10 @@ import {
   LISTENER_EDITION_TRACKING_SUFFIX,
   POOL_DETAIL_TRACKING_SUFFIX,
 } from './constants';
-import {
-  PROVISIONING_STATUS_BADGES,
-  OPERATING_STATUS_BADGES,
-} from '../../constants';
 
 export default class OctaviaLoadBalancerOverviewCtrl {
   constructor() {
     this.LISTENER_EDITION_TRACKING_SUFFIX = LISTENER_EDITION_TRACKING_SUFFIX;
     this.POOL_DETAIL_TRACKING_SUFFIX = POOL_DETAIL_TRACKING_SUFFIX;
-    this.PROVISIONING_STATUS_BADGES = PROVISIONING_STATUS_BADGES;
-    this.OPERATING_STATUS_BADGES = OPERATING_STATUS_BADGES;
   }
 }
