+++ packages/manager/modules/octavia-load-balancer/src/load-balancers/load-balancer/listeners/template.html
@@ -9,7 +9,7 @@
         <span
             data-translate="octavia_load_balancer_listeners_description_part_2"
         ></span>
-        <a href="{{ $ctrl.link }}" target="_blank" rel="noopener">
+        <a href="{{ $ctrl.discoverLink }}" target="_blank" rel="noopener">
             <span
                 data-translate="octavia_load_balancer_listeners_description_part_2_link"
             ></span>
