+++ packages/manager/modules/octavia-load-balancer/src/load-balancers/load-balancer/listeners/translations/Messages_en_GB.json
@@ -12,13 +12,6 @@
   "octavia_load_balancer_listeners_certificate": "Certificate",
   "octavia_load_balancer_listeners_port": "Port",
   "octavia_load_balancer_listeners_port_helper": "The port on which the front-end listens, must be between 1 and 65535.",
-  "octavia_load_balancer_listeners_provisioning_status": "Provisioning status",
-  "octavia_load_balancer_listeners_provisioning_status_active": "active",
-  "octavia_load_balancer_listeners_provisioning_status_creating": "creating",
-  "octavia_load_balancer_listeners_provisioning_status_deleted": "deleted",
-  "octavia_load_balancer_listeners_provisioning_status_deleting": "deleting",
-  "octavia_load_balancer_listeners_provisioning_status_error": "error",
-  "octavia_load_balancer_listeners_provisioning_status_updating": "updating",
   "octavia_load_balancer_listeners_actions_detail": "Edit",
   "octavia_load_balancer_listeners_actions_delete": "Delete",
   "octavia_load_balancer_listeners_create_success": "The new {{ listener }} listener has been created!",
