+++ packages/manager/modules/octavia-load-balancer/src/load-balancers/load-balancer/listeners/translations/Messages_fr_FR.json
@@ -12,13 +12,6 @@
   "octavia_load_balancer_listeners_certificate": "Certificat",
   "octavia_load_balancer_listeners_port": "Port",
   "octavia_load_balancer_listeners_port_helper": "Port sur lequel le front end écoute. Doit être un entier compris entre 1 et 65535.",
-  "octavia_load_balancer_listeners_provisioning_status": "Statut d'approvisionnement",
-  "octavia_load_balancer_listeners_provisioning_status_active": "actif",
-  "octavia_load_balancer_listeners_provisioning_status_creating": "en cours de création",
-  "octavia_load_balancer_listeners_provisioning_status_deleted": "supprimé",
-  "octavia_load_balancer_listeners_provisioning_status_deleting": "en cours de suppression",
-  "octavia_load_balancer_listeners_provisioning_status_error": "erreur",
-  "octavia_load_balancer_listeners_provisioning_status_updating": "en cours de mise à jour",
   "octavia_load_balancer_listeners_actions_detail": "Editer",
   "octavia_load_balancer_listeners_actions_delete": "Supprimer",
   "octavia_load_balancer_listeners_create_success": "Le nouveau listener {{ listener }} a été créé avec succès !",
