+++ packages/manager/modules/octavia-load-balancer/src/load-balancers/load-balancer/pools/component.js
@@ -3,7 +3,7 @@ import template from './template.html';
 export default {
   bindings: {
     trackBase: '<',
-    discoverOptionsLink: '<',
+    discoverLink: '<',
   },
   template,
 };
