+++ packages/manager/modules/octavia-load-balancer/src/load-balancers/load-balancer/pools/detail/members/list/component.js
@@ -9,6 +9,8 @@ export default {
     trackMemberAddInstanceAction: '<',
     goToMemberDeletion: '<',
     goToMemberEdition: '<',
+    operatingStatusBadges: '<',
+    provisioningStatusBadges: '<',
   },
   template,
 };
