+++ packages/manager/modules/octavia-load-balancer/src/load-balancers/load-balancer/pools/detail/members/list/template.html
@@ -52,6 +52,29 @@
         data-helper=":: 'octavia_load_balancer_pools_detail_members_protocol_port_tooltip' | translate"
     >
     </oui-datagrid-column>
+    <oui-datagrid-column
+        data-title=":: 'octavia_load_balancer_provisioning_status' | translate"
+        data-property="provisioningStatus"
+        data-sortable
+        data-type="options"
+    >
+        <span
+            class="oui-badge"
+            data-ng-class="$ctrl.provisioningStatusBadges[$value]"
+            data-translate="{{:: 'octavia_load_balancer_provisioning_status_' + $value }}"
+        ></span>
+    </oui-datagrid-column>
+    <oui-datagrid-column
+        data-title=":: 'octavia_load_balancer_operating_status' | translate"
+        data-property="operatingStatus"
+        data-type="options"
+    >
+        <span
+            class="oui-badge"
+            data-ng-class="$ctrl.operatingStatusBadges[$value]"
+            data-translate="{{:: 'octavia_load_balancer_operating_status_' + $value }}"
+        ></span>
+    </oui-datagrid-column>
     <oui-action-menu compact data-placement="end">
         <oui-action-menu-item data-on-click="$ctrl.goToMemberEdition($row)"
             ><span
