+++ packages/manager/modules/octavia-load-balancer/src/load-balancers/load-balancer/pools/detail/overview/component.js
@@ -1,5 +1,4 @@
 import template from './template.html';
-import controller from './controller';
 
 export default {
   bindings: {
@@ -10,7 +9,8 @@ export default {
     goToAddMemberManually: '<',
     goToAddMemberFromInstances: '<',
     goToDelete: '<',
+    operatingStatusBadges: '<',
+    provisioningStatusBadges: '<',
   },
   template,
-  controller,
 };
