--- packages/manager/modules/octavia-load-balancer/src/load-balancers/load-balancer/pools/detail/overview/controller.js
@@ -1,11 +1,0 @@
-import {
-  PROVISIONING_STATUS_BADGES,
-  OPERATING_STATUS_BADGES,
-} from '../../../constants';
-
-export default class OctaviaLoadBalancerPoolsDetailOverviewCtrl {
-  constructor() {
-    this.PROVISIONING_STATUS_BADGES = PROVISIONING_STATUS_BADGES;
-    this.OPERATING_STATUS_BADGES = OPERATING_STATUS_BADGES;
-  }
-}
