+++ packages/manager/modules/octavia-load-balancer/src/load-balancers/load-balancer/pools/list/component.js
@@ -1,4 +1,3 @@
-import controller from './controller';
 import template from './template.html';
 
 export default {
@@ -8,7 +7,8 @@ export default {
     getPoolDetailLink: '<',
     goToPoolEdition: '<',
     goToPoolDeletion: '<',
+    operatingStatusBadges: '<',
+    provisioningStatusBadges: '<',
   },
-  controller,
   template,
 };
