+++ packages/manager/modules/octavia-load-balancer/src/load-balancers/load-balancer/pools/list/template.html
@@ -48,15 +48,15 @@
         ></span>
     </oui-datagrid-column>
     <oui-datagrid-column
-        data-title=":: 'octavia_load_balancer_pools_provisioning_status' | translate"
+        data-title=":: 'octavia_load_balancer_provisioning_status' | translate"
         data-property="provisioningStatus"
         data-sortable
         data-type="options"
     >
         <span
             class="oui-badge"
-            data-ng-class="$ctrl.PROVISIONING_STATUS_BADGES[$value]"
-            data-translate="{{:: 'octavia_load_balancer_pools_provisioning_status_' + $value }}"
+            data-ng-class="$ctrl.provisioningStatusBadges[$value]"
+            data-translate="{{:: 'octavia_load_balancer_provisioning_status_' + $value }}"
         ></span>
     </oui-datagrid-column>
     <oui-datagrid-column
@@ -66,7 +66,7 @@
     >
         <span
             class="oui-badge"
-            data-ng-class="$ctrl.OPERATING_STATUS_BADGES[$value]"
+            data-ng-class="$ctrl.operatingStatusBadges[$value]"
             data-translate="{{:: 'octavia_load_balancer_operating_status_' + $value }}"
         ></span>
     </oui-datagrid-column>
