+++ packages/manager/modules/octavia-load-balancer/src/load-balancers/load-balancer/pools/list/translations/Messages_en_GB.json
@@ -4,13 +4,6 @@
   "octavia_load_balancer_pools_protocol_helper": "The protocol for which this pool and its members are listening.",
   "octavia_load_balancer_pools_algorithm": "Load Balancer algorithm",
   "octavia_load_balancer_pools_algorithm_helper": "The Load Balancer algorithm that distributes traffic between pool members.",
-  "octavia_load_balancer_pools_provisioning_status": "Provisioning status",
-  "octavia_load_balancer_pools_provisioning_status_active": "active",
-  "octavia_load_balancer_pools_provisioning_status_creating": "creating",
-  "octavia_load_balancer_pools_provisioning_status_deleted": "deleted",
-  "octavia_load_balancer_pools_provisioning_status_deleting": "deleting",
-  "octavia_load_balancer_pools_provisioning_status_error": "error",
-  "octavia_load_balancer_pools_provisioning_status_updating": "updating",
   "octavia_load_balancer_pools_actions_edit": "Edit",
   "octavia_load_balancer_pools_actions_manage_members": "Manage members",
   "octavia_load_balancer_pools_actions_delete": "Delete"
