+++ packages/manager/modules/octavia-load-balancer/src/load-balancers/load-balancer/pools/template.html
@@ -9,7 +9,7 @@
             data-translate="octavia_load_balancer_pools_description_part_2"
         ></span>
         <a
-            data-ng-href="{{:: $ctrl.discoverOptionsLink }}"
+            data-ng-href="{{:: $ctrl.discoverLink }}"
             data-track-on="click"
             data-track-name="{{:: $ctrl.trackBase + '::goto-discover-options' }}"
             data-track-type="action"
