+++ packages/manager/modules/octavia-load-balancer/src/load-balancers/load-balancer/translations/Messages_en_GB.json
@@ -4,12 +4,5 @@
   "octavia_load_balancer_pools_tab_title": "Pools",
   "octavia_load_balancer_statistics_tab_title": "Statistics",
   "octavia_load_balancer_certificates_tab_title": "Certificates",
-  "octavia_load_balancer_guides_header": "Guides",
-  "octavia_load_balancer_operating_status": "Operating status",
-  "octavia_load_balancer_operating_status_degraded": "degraded",
-  "octavia_load_balancer_operating_status_draining": "draining",
-  "octavia_load_balancer_operating_status_error": "error",
-  "octavia_load_balancer_operating_status_noMonitor": "no Health Monitor",
-  "octavia_load_balancer_operating_status_online": "online",
-  "octavia_load_balancer_operating_status_offline": "offline"
+  "octavia_load_balancer_guides_header": "Guides"
 }
