+++ packages/manager/modules/octavia-load-balancer/src/load-balancers/load-balancers.component.js
@@ -8,6 +8,7 @@ export default {
     goToDeleteLoadBalancer: '<',
     dashboardPageLink: '<',
     dashboardPageTracking: '<',
+    operatingStatusBadges: '<',
   },
   template,
 };
