+++ packages/manager/modules/octavia-load-balancer/src/translations/Messages_fr_FR.json
@@ -8,13 +8,21 @@
   "octavia_load_balancer_region": "Région",
   "octavia_load_balancer_ip": "Adresse IP privée",
   "octavia_load_balancer_creation_date": "Date de création",
-  "octavia_load_balancer_status": "Statut d'opération",
+  "octavia_load_balancer_operating_status": "Statut d'opération",
   "octavia_load_balancer_actions_detail": "Détail du Load Balancer",
   "octavia_load_balancer_actions_delete": "Supprimer",
-  "octavia_load_balancer_status_online": "Actif",
-  "octavia_load_balancer_status_offline": "Hors-ligne",
-  "octavia_load_balancer_status_degraded": "Dégradé",
-  "octavia_load_balancer_status_draining": "Purge",
-  "octavia_load_balancer_status_error": "Erreur",
+  "octavia_load_balancer_operating_status_online": "online",
+  "octavia_load_balancer_operating_status_offline": "offline",
+  "octavia_load_balancer_operating_status_degraded": "dégradé",
+  "octavia_load_balancer_operating_status_draining": "dégradé",
+  "octavia_load_balancer_operating_status_error": "erreur",
+  "octavia_load_balancer_operating_status_noMonitor": "pas de Health Monitor",
+  "octavia_load_balancer_provisioning_status": "Statut d'approvisionnement",
+  "octavia_load_balancer_provisioning_status_active": "Actif",
+  "octavia_load_balancer_provisioning_status_creating": "en cours de création",
+  "octavia_load_balancer_provisioning_status_deleted": "supprimé",
+  "octavia_load_balancer_provisioning_status_deleting": "en cours de suppression",
+  "octavia_load_balancer_provisioning_status_updating": "en cours de mise à jour",
+  "octavia_load_balancer_provisioning_status_error": "erreur",
   "octavia_load_balancer_delete_success": "Votre Load Balancer a été supprimé avec succès"
 }
